+++ Frends.Edifact.ConvertToJson/Frends.Edifact.ConvertToJson/Definitions/Options.cs
@@ -1,0 +1,25 @@
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+
+namespace Frends.Edifact.ConvertToJson.Definitions;
+
+/// <summary>
+/// Additional parameters.
+/// </summary>
+public class Options
+{
+    /// <summary>
+    /// Whether to throw an error on failure.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Overrides the error message on failure.
+    /// </summary>
+    /// <example>Order validation failed: missing customer ID</example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
+}