+++ Frends.Edifact.ConvertToJson/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.0.0] - 2026-08-24
+### Changed
+- Added Options parameter with ThrowErrorOnFailure and ErrorMessageOnFailure support.
+- Added CancellationToken parameter for cancellation support.
+- Added Success and Error properties to the Result object.
+- Upgraded target framework from .NET 6 to .NET 8.
+
 ## [1.1.0] - 2024-08-23
 ### Changed
 - Updated the Newtonsoft.Json package to version 13.0.3.